Add variable resolution
- New bind option "resolve"
- Rename test classes
- Update dependencies
