+++ src/main/java/net/jmob/guice/conf/core/impl/virtual/VirtualBean.java
@@ -16,6 +16,10 @@
 
 package net.jmob.guice.conf.core.impl.virtual;
 
+import static java.lang.Character.toUpperCase;
+import static java.lang.String.format;
+import static java.util.Optional.ofNullable;
+
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
@@ -23,10 +27,6 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
-
-import static java.lang.Character.toUpperCase;
-import static java.lang.String.format;
-import static java.util.Optional.ofNullable;
 
 final class VirtualBean implements InvocationHandler {
