+++ src/test/java/net/jmob/guice/conf/core/samples/basic/ConfigFromProperties.java
@@ -14,22 +14,24 @@
  * limitations under the License.
  */
 
-package net.jmob.guice.conf.core.samples.simple;
-
-import com.google.inject.AbstractModule;
-import net.jmob.guice.conf.core.BindConfig;
-import net.jmob.guice.conf.core.ConfigurationModule;
-import net.jmob.guice.conf.core.InjectConfig;
-import org.junit.Before;
-import org.junit.Test;
+package net.jmob.guice.conf.core.samples.basic;
 
 import static com.google.inject.Guice.createInjector;
 import static net.jmob.guice.conf.core.Syntax.PROPERTIES;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
 
-@BindConfig(value = "net/jmob/guice/conf/core/samples/sample_00", syntax = PROPERTIES)
-public class Sample00 {
+import org.junit.Before;
+import org.junit.Test;
+
+import com.google.inject.AbstractModule;
+
+import net.jmob.guice.conf.core.BindConfig;
+import net.jmob.guice.conf.core.ConfigurationModule;
+import net.jmob.guice.conf.core.InjectConfig;
+
+@BindConfig(value = "net/jmob/guice/conf/core/samples/basic", syntax = PROPERTIES)
+public class ConfigFromProperties {
 
     @InjectConfig(value = "value")
     private String value;
