+++ src/test/java/net/jmob/guice/conf/core/samples/basic/MainApp.java
@@ -1,4 +1,4 @@
-package net.jmob.guice.conf.core.samples.simple;
+package net.jmob.guice.conf.core.samples.basic;
 
 import static java.lang.String.format;
 
@@ -13,7 +13,7 @@ import net.jmob.guice.conf.core.ConfigurationModule;
 import net.jmob.guice.conf.core.InjectConfig;
 
 @BindConfig(value = "net/jmob/guice/conf/core/samples/sample_01")
-public class MyApp extends AbstractModule {
+public class MainApp extends AbstractModule {
 
     @InjectConfig
     private Optional<Integer> port;
@@ -28,8 +28,8 @@ public class MyApp extends AbstractModule {
     }
 
     public static void main(String... args)  {
-        Injector injector = Guice.createInjector(new MyApp());
-        MyApp myApp = injector.getInstance(MyApp.class);
+        Injector injector = Guice.createInjector(new MainApp());
+        MainApp myApp = injector.getInstance(MainApp.class);
         System.out.println(format("Service value : %s, port %d", myApp.config.getValue(), myApp.port.orElse(-1)));
     }
 }
