+++ src/test/resources/net/jmob/guice/conf/core/samples/basic.properties
