+++ src/test/resources/net/jmob/guice/conf/core/samples/resolved.conf
@@ -1,0 +1,3 @@
+{
+  myconfig = ${my.property}", World!"
+}
